convert boolean ui tests into unit tests
